agent/requirements.txt
strands-agents
strands-agents-tools
boto3>=1.42.3
botocore>=1.42.3
bedrock-agentcore==1.1.1
mcp
pyyaml
httpx
